refactor: split the substitution formatter to different type (#28836)
* split the substitution formatter to different type


* fix test


* fix extension


---------
