+++ source/common/formatter/substitution_formatter.h
@@ -90,28 +90,6 @@ public:
         }(params),
         ...);
   }
-
-  /**
-   * Return a FormatterProviderPtr if a built-in command is found. This method
-   * handles mapping the command name to an appropriate formatter.
-   *
-   * @param command - formatter's name.
-   * @param subcommand - optional formatter specific data.
-   * @param length - optional max length of output produced by the formatter.
-   * @return FormattterProviderPtr substitution provider for the command or nullptr
-   */
-  static FormatterProviderPtr parseBuiltinCommand(const std::string& command,
-                                                  const std::string& subcommand,
-                                                  absl::optional<size_t>& length);
-
-private:
-  using FormatterProviderCreateFunc =
-      std::function<FormatterProviderPtr(const std::string&, absl::optional<size_t>&)>;
-
-  using FormatterProviderLookupTbl =
-      absl::flat_hash_map<absl::string_view, std::pair<CommandSyntaxChecker::CommandSyntaxFlags,
-                                                       FormatterProviderCreateFunc>>;
-  static const FormatterProviderLookupTbl& getKnownFormatters();
 };
 
 /**
@@ -129,9 +107,18 @@ public:
   static const absl::optional<std::string> getHostname();
   static const std::string getHostnameOrDefault();
 
-private:
-  SubstitutionFormatUtils();
+  /**
+   * Unspecified value for protobuf.
+   */
+  static const ProtobufWkt::Value& unspecifiedValue();
 
+  /**
+   * Truncate a string to a maximum length. Do nothing if max_length is not set or
+   * max_length is greater than the length of the string.
+   */
+  static void truncate(std::string& str, absl::optional<size_t> max_length);
+
+private:
   static const std::string DEFAULT_FORMAT;
 };
 
@@ -261,447 +248,6 @@ public:
 
 private:
   const StructFormatter struct_formatter_;
-};
-
-/**
- * FormatterProvider for string literals. It ignores headers and stream info and returns string by
- * which it was initialized.
- */
-class PlainStringFormatter : public FormatterProvider {
-public:
-  PlainStringFormatter(const std::string& str);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value str_;
-};
-
-/**
- * FormatterProvider for numbers.
- */
-class PlainNumberFormatter : public FormatterProvider {
-public:
-  PlainNumberFormatter(double num);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value num_;
-};
-
-/**
- * FormatterProvider for local_reply_body. It returns the string from `local_reply_body` argument.
- */
-class LocalReplyBodyFormatter : public FormatterProvider {
-public:
-  LocalReplyBodyFormatter() = default;
-
-  // Formatter::format
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view local_reply_body,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view local_reply_body,
-                                 AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for access log type. It returns the string from `access_log_type` argument.
- */
-class AccessLogTypeFormatter : public FormatterProvider {
-public:
-  AccessLogTypeFormatter() = default;
-
-  // Formatter::format
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view local_reply_body,
-                                     AccessLog::AccessLogType access_log_type) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view local_reply_body,
-                                 AccessLog::AccessLogType access_log_type) const override;
-};
-
-class HeaderFormatter {
-public:
-  HeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                  absl::optional<size_t> max_length);
-
-protected:
-  absl::optional<std::string> format(const Http::HeaderMap& headers) const;
-  ProtobufWkt::Value formatValue(const Http::HeaderMap& headers) const;
-
-private:
-  const Http::HeaderEntry* findHeader(const Http::HeaderMap& headers) const;
-
-  Http::LowerCaseString main_header_;
-  Http::LowerCaseString alternative_header_;
-  absl::optional<size_t> max_length_;
-};
-
-/**
- * FormatterProvider for headers byte size.
- */
-class HeadersByteSizeFormatter : public FormatterProvider {
-public:
-  // TODO(taoxuy): Add RequestTrailers here.
-  enum class HeaderType { RequestHeaders, ResponseHeaders, ResponseTrailers };
-
-  HeadersByteSizeFormatter(const HeaderType header_type);
-
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap& request_headers,
-                                 const Http::ResponseHeaderMap& response_headers,
-                                 const Http::ResponseTrailerMap& response_trailers,
-                                 const StreamInfo::StreamInfo&, absl::string_view,
-                                 AccessLog::AccessLogType) const override;
-
-private:
-  uint64_t extractHeadersByteSize(const Http::RequestHeaderMap& request_headers,
-                                  const Http::ResponseHeaderMap& response_headers,
-                                  const Http::ResponseTrailerMap& response_trailers) const;
-  HeaderType header_type_;
-};
-
-/**
- * FormatterProvider for request headers.
- */
-class RequestHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  RequestHeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                         absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for response headers.
- */
-class ResponseHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  ResponseHeaderFormatter(const std::string& main_header, const std::string& alternative_header,
-                          absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-/**
- * FormatterProvider for response trailers.
- */
-class ResponseTrailerFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  ResponseTrailerFormatter(const std::string& main_header, const std::string& alternative_header,
-                           absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-};
-
-class GrpcStatusFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  enum Format {
-    CamelString,
-    SnakeString,
-    Number,
-  };
-
-  GrpcStatusFormatter(const std::string& main_header, const std::string& alternative_header,
-                      absl::optional<size_t> max_length, Format format);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&,
-                                     const Http::ResponseHeaderMap& response_headers,
-                                     const Http::ResponseTrailerMap& response_trailers,
-                                     const StreamInfo::StreamInfo&, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-  static Format parseFormat(absl::string_view format);
-
-private:
-  const Format format_;
-};
-
-/**
- * FormatterProvider based on StreamInfo fields.
- */
-class StreamInfoFormatter : public FormatterProvider {
-public:
-  StreamInfoFormatter(const std::string&, const std::string& = "",
-                      const absl::optional<size_t>& = absl::nullopt);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-  class FieldExtractor {
-  public:
-    virtual ~FieldExtractor() = default;
-
-    virtual absl::optional<std::string> extract(const StreamInfo::StreamInfo&) const PURE;
-    virtual ProtobufWkt::Value extractValue(const StreamInfo::StreamInfo&) const PURE;
-  };
-  using FieldExtractorPtr = std::unique_ptr<FieldExtractor>;
-  using FieldExtractorCreateFunc =
-      std::function<FieldExtractorPtr(const std::string&, const absl::optional<size_t>&)>;
-
-  enum class StreamInfoAddressFieldExtractionType { WithPort, WithoutPort, JustPort };
-  StreamInfoFormatter(FieldExtractorPtr field_extractor) {
-    field_extractor_ = std::move(field_extractor);
-  }
-
-private:
-  FieldExtractorPtr field_extractor_;
-
-  using FieldExtractorLookupTbl =
-      absl::flat_hash_map<absl::string_view,
-                          std::pair<CommandSyntaxChecker::CommandSyntaxFlags,
-                                    StreamInfoFormatter::FieldExtractorCreateFunc>>;
-  static const FieldExtractorLookupTbl& getKnownFieldExtractors();
-};
-
-/**
- * Base formatter for formatting Metadata objects
- */
-
-class MetadataFormatter : public FormatterProvider {
-public:
-  using GetMetadataFunction =
-      std::function<const envoy::config::core::v3::Metadata*(const StreamInfo::StreamInfo&)>;
-  MetadataFormatter(const std::string& filter_namespace, const std::vector<std::string>& path,
-                    absl::optional<size_t> max_length, GetMetadataFunction get);
-
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&,
-                                     const StreamInfo::StreamInfo& stream_info, absl::string_view,
-                                     AccessLog::AccessLogType) const override;
-
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&,
-                                 const StreamInfo::StreamInfo& stream_info, absl::string_view,
-                                 AccessLog::AccessLogType) const override;
-
-protected:
-  absl::optional<std::string>
-  formatMetadata(const envoy::config::core::v3::Metadata& metadata) const;
-  ProtobufWkt::Value formatMetadataValue(const envoy::config::core::v3::Metadata& metadata) const;
-
-private:
-  std::string filter_namespace_;
-  std::vector<std::string> path_;
-  absl::optional<size_t> max_length_;
-  GetMetadataFunction get_func_;
-};
-
-/**
- * FormatterProvider for DynamicMetadata from StreamInfo.
- */
-class DynamicMetadataFormatter : public MetadataFormatter {
-public:
-  DynamicMetadataFormatter(const std::string& filter_namespace,
-                           const std::vector<std::string>& path, absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for ClusterMetadata from StreamInfo.
- */
-class ClusterMetadataFormatter : public MetadataFormatter {
-public:
-  ClusterMetadataFormatter(const std::string& filter_namespace,
-                           const std::vector<std::string>& path, absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for UpstreamHostMetadata from StreamInfo.
- */
-class UpstreamHostMetadataFormatter : public MetadataFormatter {
-public:
-  UpstreamHostMetadataFormatter(const std::string& filter_namespace,
-                                const std::vector<std::string>& path,
-                                absl::optional<size_t> max_length);
-};
-
-/**
- * FormatterProvider for FilterState from StreamInfo.
- */
-class FilterStateFormatter : public FormatterProvider {
-public:
-  static std::unique_ptr<FilterStateFormatter>
-  create(const std::string& format, const absl::optional<size_t>& max_length, bool is_upstream);
-
-  FilterStateFormatter(const std::string& key, absl::optional<size_t> max_length,
-                       bool serialize_as_string, bool is_upstream = false);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  const Envoy::StreamInfo::FilterState::Object*
-  filterState(const StreamInfo::StreamInfo& stream_info) const;
-
-  std::string key_;
-  absl::optional<size_t> max_length_;
-
-  bool serialize_as_string_;
-  const bool is_upstream_;
-};
-
-/**
- * Base FormatterProvider for system times from StreamInfo.
- */
-class SystemTimeFormatter : public FormatterProvider {
-public:
-  using TimeFieldExtractor =
-      std::function<absl::optional<SystemTime>(const StreamInfo::StreamInfo& stream_info)>;
-  using TimeFieldExtractorPtr = std::unique_ptr<TimeFieldExtractor>;
-
-  SystemTimeFormatter(const std::string& format, TimeFieldExtractorPtr f);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  const Envoy::DateFormatter date_formatter_;
-  const TimeFieldExtractorPtr time_field_extractor_;
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for request start time from StreamInfo.
- */
-class StartTimeFormatter : public SystemTimeFormatter {
-public:
-  StartTimeFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for downstream cert start time from the StreamInfo's
- * ConnectionInfo.
- */
-class DownstreamPeerCertVStartFormatter : public SystemTimeFormatter {
-public:
-  DownstreamPeerCertVStartFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for downstream cert end time from the StreamInfo's
- * ConnectionInfo.
- */
-class DownstreamPeerCertVEndFormatter : public SystemTimeFormatter {
-public:
-  DownstreamPeerCertVEndFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for upstream cert start time from the StreamInfo's
- * upstreamInfo.
- */
-class UpstreamPeerCertVStartFormatter : public SystemTimeFormatter {
-public:
-  UpstreamPeerCertVStartFormatter(const std::string& format);
-};
-
-/**
- * SystemTimeFormatter (FormatterProvider) for upstream cert end time from the StreamInfo's
- * upstreamInfo.
- */
-class UpstreamPeerCertVEndFormatter : public SystemTimeFormatter {
-public:
-  UpstreamPeerCertVEndFormatter(const std::string& format);
-};
-
-/**
- * FormatterProvider for environment. If no valid environment value then
- */
-class EnvironmentFormatter : public FormatterProvider {
-public:
-  EnvironmentFormatter(const std::string& key, absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
-
-private:
-  ProtobufWkt::Value str_;
-};
-
-/**
- * FormatterProvider for request headers from StreamInfo (rather than the request_headers param).
- * Purely for testing.
- */
-class StreamInfoRequestHeaderFormatter : public FormatterProvider, HeaderFormatter {
-public:
-  StreamInfoRequestHeaderFormatter(const std::string& main_header,
-                                   const std::string& alternative_header,
-                                   absl::optional<size_t> max_length);
-
-  // FormatterProvider
-  absl::optional<std::string> format(const Http::RequestHeaderMap& request_headers,
-                                     const Http::ResponseHeaderMap&,
-                                     const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                     absl::string_view, AccessLog::AccessLogType) const override;
-  ProtobufWkt::Value formatValue(const Http::RequestHeaderMap&, const Http::ResponseHeaderMap&,
-                                 const Http::ResponseTrailerMap&, const StreamInfo::StreamInfo&,
-                                 absl::string_view, AccessLog::AccessLogType) const override;
 };
 
 } // namespace Formatter
