+++ source/extensions/formatter/metadata/metadata.h
@@ -5,6 +5,7 @@
 #include "envoy/config/typed_config.h"
 #include "envoy/registry/registry.h"
 
+#include "source/common/formatter/stream_info_formatter.h"
 #include "source/common/formatter/substitution_formatter.h"
 
 namespace Envoy {
@@ -22,7 +23,7 @@ public:
 private:
   // Map used to dispatch types of metadata to individual handlers which will
   // access required metadata object.
-  using FormatterProviderFunc = std::function<::Envoy::Formatter::FormatterProviderPtr(
+  using FormatterProviderFunc = std::function<::Envoy::Formatter::StreamInfoFormatterProviderPtr(
       const std::string& filter_namespace, const std::vector<std::string>& path,
       absl::optional<size_t> max_length)>;
   std::map<std::string, FormatterProviderFunc> metadata_formatter_providers_;
