+++ test/common/formatter/substitution_formatter_test.cc
@@ -9,6 +9,8 @@
 
 #include "source/common/common/logger.h"
 #include "source/common/common/utility.h"
+#include "source/common/formatter/http_specific_formatter.h"
+#include "source/common/formatter/stream_info_formatter.h"
 #include "source/common/formatter/substitution_formatter.h"
 #include "source/common/http/header_map_impl.h"
 #include "source/common/json/json_loader.h"
@@ -2392,7 +2394,7 @@ void populateMetadataTestData(envoy::config::core::v3::Metadata& metadata) {
   (*metadata.mutable_filter_metadata())["com.test"] = struct_obj;
 }
 
-TEST(SubstitutionFormatterTest, DynamicMetadataFormatter) {
+TEST(SubstitutionFormatterTest, DynamicMetadataFieldExtractor) {
   envoy::config::core::v3::Metadata metadata;
   populateMetadataTestData(metadata);
   NiceMock<StreamInfo::MockStreamInfo> stream_info;
@@ -2405,87 +2407,60 @@ TEST(SubstitutionFormatterTest, DynamicMetadataFormatter) {
 
   {
     DynamicMetadataFormatter formatter("com.test", {}, absl::optional<size_t>());
-    std::string val = formatter
-                          .format(request_headers, response_headers, response_trailers, stream_info,
-                                  body, AccessLog::AccessLogType::NotSet)
-                          .value();
+    std::string val = formatter.format(stream_info).value();
     EXPECT_TRUE(val.find("\"test_key\":\"test_value\"") != std::string::npos);
     EXPECT_TRUE(val.find("\"test_obj\":{\"inner_key\":\"inner_value\"}") != std::string::npos);
 
     ProtobufWkt::Value expected_val;
     expected_val.mutable_struct_value()->CopyFrom(metadata.filter_metadata().at("com.test"));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(expected_val));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(expected_val));
   }
   {
     DynamicMetadataFormatter formatter("com.test", {"test_key"}, absl::optional<size_t>());
-    EXPECT_EQ("test_value", formatter.format(request_headers, response_headers, response_trailers,
-                                             stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::stringValue("test_value")));
+    EXPECT_EQ("test_value", formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::stringValue("test_value")));
   }
   {
     DynamicMetadataFormatter formatter("com.test", {"test_obj"}, absl::optional<size_t>());
-    EXPECT_EQ("{\"inner_key\":\"inner_value\"}",
-              formatter.format(request_headers, response_headers, response_trailers, stream_info,
-                               body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("{\"inner_key\":\"inner_value\"}", formatter.format(stream_info));
 
     ProtobufWkt::Value expected_val;
     (*expected_val.mutable_struct_value()->mutable_fields())["inner_key"] =
         ValueUtil::stringValue("inner_value");
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(expected_val));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(expected_val));
   }
   {
     DynamicMetadataFormatter formatter("com.test", {"test_obj", "inner_key"},
                                        absl::optional<size_t>());
-    EXPECT_EQ("inner_value", formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::stringValue("inner_value")));
+    EXPECT_EQ("inner_value", formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::stringValue("inner_value")));
   }
 
   // not found cases
   {
     DynamicMetadataFormatter formatter("com.notfound", {}, absl::optional<size_t>());
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   {
     DynamicMetadataFormatter formatter("com.test", {"notfound"}, absl::optional<size_t>());
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   {
     DynamicMetadataFormatter formatter("com.test", {"test_obj", "notfound"},
                                        absl::optional<size_t>());
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
 
   // size limit
   {
     DynamicMetadataFormatter formatter("com.test", {"test_key"}, absl::optional<size_t>(5));
-    EXPECT_EQ("test_", formatter.format(request_headers, response_headers, response_trailers,
-                                        stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("test_", formatter.format(stream_info));
 
     // N.B. Does not truncate.
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::stringValue("test_value")));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::stringValue("test_value")));
   }
 
   {
@@ -2496,9 +2471,7 @@ TEST(SubstitutionFormatterTest, DynamicMetadataFormatter) {
     (*metadata.mutable_filter_metadata())["com.test"] = struct_obj;
 
     DynamicMetadataFormatter formatter("com.test", {"nan_val"}, absl::optional<size_t>());
-    absl::optional<std::string> value =
-        formatter.format(request_headers, response_headers, response_trailers, stream_info, body,
-                         AccessLog::AccessLogType::NotSet);
+    absl::optional<std::string> value = formatter.format(stream_info);
     EXPECT_EQ("google.protobuf.Value cannot encode double values for nan, because it would be "
               "parsed as a string",
               value.value());
@@ -2512,9 +2485,7 @@ TEST(SubstitutionFormatterTest, DynamicMetadataFormatter) {
     (*metadata.mutable_filter_metadata())["com.test"] = struct_obj;
 
     DynamicMetadataFormatter formatter("com.test", {"inf_val"}, absl::optional<size_t>());
-    absl::optional<std::string> value =
-        formatter.format(request_headers, response_headers, response_trailers, stream_info, body,
-                         AccessLog::AccessLogType::NotSet);
+    absl::optional<std::string> value = formatter.format(stream_info);
     EXPECT_EQ("google.protobuf.Value cannot encode double values for infinity, because it would be "
               "parsed as a string",
               value.value());
@@ -2550,101 +2521,75 @@ TEST(SubstitutionFormatterTest, FilterStateFormatter) {
   {
     FilterStateFormatter formatter("key", absl::optional<size_t>(), false);
 
-    EXPECT_EQ("\"test_value\"",
-              formatter.format(request_headers, response_headers, response_trailers, stream_info,
-                               body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::stringValue("test_value")));
+    EXPECT_EQ("\"test_value\"", formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::stringValue("test_value")));
   }
   {
     FilterStateFormatter formatter("key-struct", absl::optional<size_t>(), false);
 
-    EXPECT_EQ("{\"inner_key\":\"inner_value\"}",
-              formatter.format(request_headers, response_headers, response_trailers, stream_info,
-                               body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("{\"inner_key\":\"inner_value\"}", formatter.format(stream_info));
 
     ProtobufWkt::Value expected;
     (*expected.mutable_struct_value()->mutable_fields())["inner_key"] =
         ValueUtil::stringValue("inner_value");
 
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(expected));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(expected));
   }
 
   // not found case
   {
     FilterStateFormatter formatter("key-not-found", absl::optional<size_t>(), false);
 
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
 
   // no serialization case
   {
     FilterStateFormatter formatter("key-no-serialization", absl::optional<size_t>(), false);
 
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
 
   // serialization error case
   {
     FilterStateFormatter formatter("key-serialization-error", absl::optional<size_t>(), false);
 
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
 
   // size limit
   {
     FilterStateFormatter formatter("key", absl::optional<size_t>(5), false);
 
-    EXPECT_EQ("\"test", formatter.format(request_headers, response_headers, response_trailers,
-                                         stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("\"test", formatter.format(stream_info));
 
     // N.B. Does not truncate.
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::stringValue("test_value")));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::stringValue("test_value")));
   }
 
   // serializeAsString case
   {
     FilterStateFormatter formatter("test_key", absl::optional<size_t>(), true);
 
-    EXPECT_EQ("test_value By PLAIN",
-              formatter.format(request_headers, response_headers, response_trailers, stream_info,
-                               body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("test_value By PLAIN", formatter.format(stream_info));
   }
 
   // size limit for serializeAsString
   {
     FilterStateFormatter formatter("test_key", absl::optional<size_t>(10), true);
 
-    EXPECT_EQ("test_value", formatter.format(request_headers, response_headers, response_trailers,
-                                             stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("test_value", formatter.format(stream_info));
   }
 
   // no serialization case for serializeAsString
   {
     FilterStateFormatter formatter("key-no-serialization", absl::optional<size_t>(), true);
 
-    EXPECT_EQ(absl::nullopt, formatter.format(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(formatter.formatValue(request_headers, response_headers, response_trailers,
-                                      stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, formatter.format(stream_info));
+    EXPECT_THAT(formatter.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
 }
 
@@ -2659,12 +2604,8 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVStartFormatter) {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     stream_info.downstream_connection_info_provider_->setSslConnection(nullptr);
     DownstreamPeerCertVStartFormatter cert_start_formart("DOWNSTREAM_PEER_CERT_V_START(%Y/%m/%d)");
-    EXPECT_EQ(absl::nullopt,
-              cert_start_formart.format(request_headers, response_headers, response_trailers,
-                                        stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_start_formart.formatValue(request_headers, response_headers, response_trailers,
-                                               stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_start_formart.format(stream_info));
+    EXPECT_THAT(cert_start_formart.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No validFromPeerCertificate
   {
@@ -2673,12 +2614,8 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVStartFormatter) {
     auto connection_info = std::make_shared<Ssl::MockConnectionInfo>();
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(absl::nullopt));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ(absl::nullopt,
-              cert_start_formart.format(request_headers, response_headers, response_trailers,
-                                        stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_start_formart.formatValue(request_headers, response_headers, response_trailers,
-                                               stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_start_formart.format(stream_info));
+    EXPECT_THAT(cert_start_formart.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // Default format string
   {
@@ -2689,9 +2626,7 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVStartFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time),
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time), cert_start_format.format(stream_info));
   }
   // Custom format string
   {
@@ -2702,9 +2637,7 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVStartFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ("Mar 28 23:35:58 2018 UTC",
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("Mar 28 23:35:58 2018 UTC", cert_start_format.format(stream_info));
   }
 }
 
@@ -2719,12 +2652,8 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVEndFormatter) {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     stream_info.downstream_connection_info_provider_->setSslConnection(nullptr);
     DownstreamPeerCertVEndFormatter cert_end_format("%Y/%m/%d");
-    EXPECT_EQ(absl::nullopt,
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_end_format.formatValue(request_headers, response_headers, response_trailers,
-                                            stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_end_format.format(stream_info));
+    EXPECT_THAT(cert_end_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No expirationPeerCertificate
   {
@@ -2734,12 +2663,8 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVEndFormatter) {
     EXPECT_CALL(*connection_info, expirationPeerCertificate())
         .WillRepeatedly(Return(absl::nullopt));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ(absl::nullopt,
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_end_format.formatValue(request_headers, response_headers, response_trailers,
-                                            stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_end_format.format(stream_info));
+    EXPECT_THAT(cert_end_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // Default format string
   {
@@ -2750,9 +2675,7 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVEndFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, expirationPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time),
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time), cert_end_format.format(stream_info));
   }
   // Custom format string
   {
@@ -2763,9 +2686,7 @@ TEST(SubstitutionFormatterTest, DownstreamPeerCertVEndFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, expirationPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.downstream_connection_info_provider_->setSslConnection(connection_info);
-    EXPECT_EQ("Mar 28 23:35:58 2018 UTC",
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("Mar 28 23:35:58 2018 UTC", cert_end_format.format(stream_info));
   }
 }
 
@@ -2780,24 +2701,16 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVStartFormatter) {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     EXPECT_CALL(stream_info, upstreamInfo()).WillRepeatedly(Return(nullptr));
     UpstreamPeerCertVStartFormatter cert_start_format("%Y/%m/%d");
-    EXPECT_EQ(absl::nullopt,
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_start_format.formatValue(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_start_format.format(stream_info));
+    EXPECT_THAT(cert_start_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No upstreamSslConnection
   {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     stream_info.upstreamInfo()->setUpstreamSslConnection(nullptr);
     DownstreamPeerCertVStartFormatter cert_start_format("UPSTREAM_PEER_CERT_V_START(%Y/%m/%d)");
-    EXPECT_EQ(absl::nullopt,
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_start_format.formatValue(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_start_format.format(stream_info));
+    EXPECT_THAT(cert_start_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No validFromPeerCertificate
   {
@@ -2806,12 +2719,8 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVStartFormatter) {
     auto connection_info = std::make_shared<Ssl::MockConnectionInfo>();
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(absl::nullopt));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ(absl::nullopt,
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_start_format.formatValue(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_start_format.format(stream_info));
+    EXPECT_THAT(cert_start_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // Default format string
   {
@@ -2822,9 +2731,7 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVStartFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time),
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time), cert_start_format.format(stream_info));
   }
   // Custom format string
   {
@@ -2835,9 +2742,7 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVStartFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, validFromPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ("Mar 28 23:35:58 2018 UTC",
-              cert_start_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("Mar 28 23:35:58 2018 UTC", cert_start_format.format(stream_info));
   }
 }
 
@@ -2852,24 +2757,16 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVEndFormatter) {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     EXPECT_CALL(stream_info, upstreamInfo()).WillRepeatedly(Return(nullptr));
     UpstreamPeerCertVEndFormatter cert_end_format("%Y/%m/%d");
-    EXPECT_EQ(absl::nullopt,
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_end_format.formatValue(request_headers, response_headers, response_trailers,
-                                            stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_end_format.format(stream_info));
+    EXPECT_THAT(cert_end_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No upstreamSslConnection
   {
     NiceMock<StreamInfo::MockStreamInfo> stream_info;
     stream_info.upstreamInfo()->setUpstreamSslConnection(nullptr);
     UpstreamPeerCertVEndFormatter cert_end_format("%Y/%m/%d");
-    EXPECT_EQ(absl::nullopt,
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_end_format.formatValue(request_headers, response_headers, response_trailers,
-                                            stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_end_format.format(stream_info));
+    EXPECT_THAT(cert_end_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // No expirationPeerCertificate
   {
@@ -2879,12 +2776,8 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVEndFormatter) {
     EXPECT_CALL(*connection_info, expirationPeerCertificate())
         .WillRepeatedly(Return(absl::nullopt));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ(absl::nullopt,
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(cert_end_format.formatValue(request_headers, response_headers, response_trailers,
-                                            stream_info, body, AccessLog::AccessLogType::NotSet),
-                ProtoEq(ValueUtil::nullValue()));
+    EXPECT_EQ(absl::nullopt, cert_end_format.format(stream_info));
+    EXPECT_THAT(cert_end_format.formatValue(stream_info), ProtoEq(ValueUtil::nullValue()));
   }
   // Default format string
   {
@@ -2895,9 +2788,7 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVEndFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, expirationPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time),
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time), cert_end_format.format(stream_info));
   }
   // Custom format string
   {
@@ -2908,9 +2799,7 @@ TEST(SubstitutionFormatterTest, UpstreamPeerCertVEndFormatter) {
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(*connection_info, expirationPeerCertificate()).WillRepeatedly(Return(time));
     stream_info.upstreamInfo()->setUpstreamSslConnection(connection_info);
-    EXPECT_EQ("Mar 28 23:35:58 2018 UTC",
-              cert_end_format.format(request_headers, response_headers, response_trailers,
-                                     stream_info, body, AccessLog::AccessLogType::NotSet));
+    EXPECT_EQ("Mar 28 23:35:58 2018 UTC", cert_end_format.format(stream_info));
   }
 }
 
@@ -2926,11 +2815,8 @@ TEST(SubstitutionFormatterTest, StartTimeFormatter) {
     time_t test_epoch = 1522280158;
     SystemTime time = std::chrono::system_clock::from_time_t(test_epoch);
     EXPECT_CALL(stream_info, startTime()).WillRepeatedly(Return(time));
-    EXPECT_EQ("2018/03/28",
-              start_time_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(start_time_format.formatValue(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet),
+    EXPECT_EQ("2018/03/28", start_time_format.format(stream_info));
+    EXPECT_THAT(start_time_format.formatValue(stream_info),
                 ProtoEq(ValueUtil::stringValue("2018/03/28")));
   }
 
@@ -2938,11 +2824,8 @@ TEST(SubstitutionFormatterTest, StartTimeFormatter) {
     StartTimeFormatter start_time_format("");
     SystemTime time;
     EXPECT_CALL(stream_info, startTime()).WillRepeatedly(Return(time));
-    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time),
-              start_time_format.format(request_headers, response_headers, response_trailers,
-                                       stream_info, body, AccessLog::AccessLogType::NotSet));
-    EXPECT_THAT(start_time_format.formatValue(request_headers, response_headers, response_trailers,
-                                              stream_info, body, AccessLog::AccessLogType::NotSet),
+    EXPECT_EQ(AccessLogDateTimeFormatter::fromTime(time), start_time_format.format(stream_info));
+    EXPECT_THAT(start_time_format.formatValue(stream_info),
                 ProtoEq(ValueUtil::stringValue(AccessLogDateTimeFormatter::fromTime(time))));
   }
 }
